+++ src/Service/PaymentMethodService.php
@@ -390,12 +390,10 @@ class PaymentMethodService
                 $payment['issuer'] = $issuer;
             }
 
-            $orderData->setPayment($payment);
-
             if ($molPaymentMethod->id_method === PaymentMethod::CREDITCARD) {
                 $molCustomer = $this->handleCustomerInfo($cart->id_customer, $saveCard, $useSavedCard);
                 if ($molCustomer) {
-                    $orderData->setCustomerId($molCustomer->customer_id);
+                    $payment['customerId'] = $molCustomer->customer_id;
                 }
             }