+++ changelog.md
@@ -2,6 +2,7 @@
 
 # Changelog #
 ## Changes in release 5.2.0 ##
++ Fixed issue where customer id was not added when using saved credit card for order API.
 + Added QR code option for bancontact payment on live mode
 + Added in3 payment method restriction for using only order API.